base_distanceをfloat型に変える

+++ lib/minimum/controller/pid_follow_controller.h
@@ -10,7 +10,7 @@ class PidFollowController : public CarController {
 public:
     explicit PidFollowController(
         FollowCar *car,
-        int base_distance,
+        float base_distance,
         int base_speed,
         int ir_sensor_diff,
         int max_manipulation_dist,
@@ -28,7 +28,7 @@ public:
 protected:
     FollowCar *car_;
     SonicSensorUpdater *sensor_updater_;
-    int base_distance_;
+    float base_distance_;
     int base_speed_;
     int lr_sensor_diff_; // 左右のセンサの個体差
     int max_manipulation_dist_;
